Convert into functional component and use react over preact
